+++ prisma/schema.prisma
@@ -33,10 +33,9 @@ model ProblemSession {
   lectureId String
   problemId String
 
-  problemVariablesSeed   String
-  currentProblemType     String
-  currentTraceItemIndex  Int
-  previousTraceItemIndex Int
+  problemVariablesSeed String
+  problemType          String
+  traceItemIndex       Int
 
   elapsedMilliseconds Int       @default(0)
   completedAt         DateTime?
